Quality Updates
Allowed Calculations to the side and complexity of the project so that it can easily be adjusted depending on the amount of colours you wish to display. I have left the default as 32^3.

I have removed some classes and converted them to interfaces to assist in performance optimisation.

Converted nested for loop into inbuilt JavaScript functions where possible to assit in performance optimisation.